+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidyr (development version)
 
+* New `unnest_tree()` to unnest a recursive data frame (@mgirlich, #1384).
+
 * `pivot_longer()` now throws a slightly better error message when
   `values_ptypes` or `names_ptypes` is provided and the coercion can't be made
   (#1364).